+++ examples/classic-remix/remix.config.js
@@ -20,5 +20,7 @@ module.exports = {
     v3_fetcherPersist: true,
     v3_relativeSplatpath: true,
     v3_throwAbortReason: true,
+    v3_lazyRouteDiscovery: true,
+    v3_singleFetch: true,
   },
 };