+++ examples/express/app/root.tsx
@@ -1,5 +1,4 @@
 import {
-  defer,
   type LinksFunction,
   type LoaderFunctionArgs,
 } from '@remix-run/node';
@@ -15,9 +14,18 @@ import type {Cart, Shop} from '@shopify/hydrogen/storefront-api-types';
 import styles from './styles/app.css?url';
 import {useNonce} from '@shopify/hydrogen';
 
+/**
+ * The main and reset stylesheets are added in the Layout component
+ * to prevent a bug in development HMR updates.
+ *
+ * This avoids the "failed to execute 'insertBefore' on 'Node'" error
+ * that occurs after editing and navigating to another page.
+ *
+ * It's a temporary fix until the issue is resolved.
+ * https://github.com/remix-run/remix/issues/9242
+ */
 export const links: LinksFunction = () => {
   return [
-    {rel: 'stylesheet', href: styles},
     {
       rel: 'preconnect',
       href: 'https://cdn.shopify.com',
@@ -57,11 +65,11 @@ export async function loader({context}: LoaderFunctionArgs) {
     await context.storefront.query<{shop: Shop}>(LAYOUT_QUERY),
   ]);
 
-  return defer({
+  return {
     isLoggedIn: Boolean(customerAccessToken),
     cart,
     layout,
-  });
+  };
 }
 
 export function Layout({children}: {children?: React.ReactNode}) {
@@ -75,6 +83,7 @@ export function Layout({children}: {children?: React.ReactNode}) {
       <head>
         <meta charSet="utf-8" />
         <meta name="viewport" content="width=device-width,initial-scale=1" />
+        <link rel="stylesheet" href={styles}></link>
         <Meta />
         <Links />
       </head>