+++ packages/hydrogen/CHANGELOG.md
@@ -1,5 +1,51 @@
 # @shopify/hydrogen
 
+## 2025.1.0
+
+### Patch Changes
+
+- Bump vite, Remix versions and tailwind v4 alpha to beta ([#2696](https://github.com/Shopify/hydrogen/pull/2696)) by [@wizardlyhel](https://github.com/wizardlyhel)
+
+- Remove deprecated `customerAccountUrl` ([#2730](https://github.com/Shopify/hydrogen/pull/2730)) by [@wizardlyhel](https://github.com/wizardlyhel)
+
+- Fix `getProductOptions` crashing when one of variants returns a null `firstSelectableVariant`. ([#2704](https://github.com/Shopify/hydrogen/pull/2704)) by [@wizardlyhel](https://github.com/wizardlyhel)
+
+- Bump SFAPI to 2025-01 ([#2715](https://github.com/Shopify/hydrogen/pull/2715)) by [@rbshop](https://github.com/rbshop)
+
+- Updated dependencies [[`fdab06f5`](https://github.com/Shopify/hydrogen/commit/fdab06f5d34076b526d406698bdf6fca6787660b), [`650d57b3`](https://github.com/Shopify/hydrogen/commit/650d57b3e07125661e23900e73c0bb3027ddbcde), [`064de138`](https://github.com/Shopify/hydrogen/commit/064de13890c68cabb1c3fdbe7f77409a0cf1c384), [`08b7fa5b`](https://github.com/Shopify/hydrogen/commit/08b7fa5bb99632e4707d0dec4e06f6d16b2816d0)]:
+  - @shopify/hydrogen-react@2025.1.0
+
+## 2024.10.1
+
+### Patch Changes
+
+- Added namespace support to prevent conflicts when using multiple Pagination components: ([#2649](https://github.com/Shopify/hydrogen/pull/2649)) by [@scottdixon](https://github.com/scottdixon)
+
+  - New optional `namespace` prop for the `<Pagination/>` component
+  - New optional `namespace` option for `getPaginationVariables()` utility
+  - When specified, pagination URL parameters are prefixed with the namespace (e.g., `products_cursor` instead of `cursor`)
+  - Maintains backwards compatibility when no namespace is provided
+
+- Introduce `getProductOptions`, `getAdjacentAndFirstAvailableVariants`, `useSelectedOptionInUrlParam`, and `mapSelectedProductOptionToObject` to support combined listing products and products with 2000 variants limit. ([#2659](https://github.com/Shopify/hydrogen/pull/2659)) by [@wizardlyhel](https://github.com/wizardlyhel)
+
+- Add params to override the login and authorize paths: ([#2648](https://github.com/Shopify/hydrogen/pull/2648)) by [@blittle](https://github.com/blittle)
+
+  ```ts
+  const hydrogenContext = createHydrogenContext({
+    // ...
+    customerAccount: {
+      loginPath = '/account/login',
+      authorizePath = '/account/authorize',
+      defaultRedirectPath = '/account',
+    },
+  });
+  ```
+
+- Add `selectedVariant` prop to the `VariantSelector` to use for the initial state if no URL parameters are set ([#2643](https://github.com/Shopify/hydrogen/pull/2643)) by [@scottdixon](https://github.com/scottdixon)
+
+- Updated dependencies [[`a57d5267`](https://github.com/Shopify/hydrogen/commit/a57d5267daa2f22fe1a426fb9f62c242957f95b6)]:
+  - @shopify/hydrogen-react@2024.10.1
+
 ## 2024.10.0
 
 ### Patch Changes